pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-dance>=7.1.0",
    "flask-login>=0.6.3",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "kagglehub>=0.3.12",
    "keras>=2.14.0",
    "matplotlib>=3.10.1",
    "numpy>=2.2.5",
    "oauthlib>=3.2.2",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
    "tensorflow>=2.14.0",
    "werkzeug>=3.1.3",
]
